break hook out of component into it's own generic component
